fix: 修复 clippy lint 警告并优化代码
- 移除 xu_runtime/src/lib.rs 中大部分全局 #![allow(...)] 指令
- 修复 clone_on_copy: 对 Copy 类型使用 * 而非 .clone()
- 修复 map_or -> is_some_and/is_none_or 更惯用的写法
- 修复 unwrap_or_else -> unwrap_or_default
- 修复 hash_one 替代手动创建 hasher
- 修复 first() 替代 get(0)
- 修复 copied() 替代 cloned() 用于 Copy 类型
- 修复 &[T] 替代 &Vec<T> 用于函数参数
- 修复 if_same_then_else: 合并相同分支
- 修复多余的类型转换 (as i64, as u64, .into())
- 修复 length comparison to one
- 保留 DictKey::eq 的显式短路返回以保持性能

+++ crates/xu_runtime/src/builtins/core.rs
@@ -14,7 +14,7 @@ pub fn builtin_print(rt: &mut Runtime, args: &[Value]) -> Result<Value, String> 
 pub fn builtin_gen_id(rt: &mut Runtime, _args: &[Value]) -> Result<Value, String> {
     let id = rt.types.next_id;
     rt.types.next_id = rt.types.next_id.saturating_add(1);
-    Ok(Value::from_i64(id as i64))
+    Ok(Value::from_i64(id))
 }
 
 pub fn builtin_gc(rt: &mut Runtime, _args: &[Value]) -> Result<Value, String> {
